refactor: move normalizePathSeparators to utils/path
