+++ packages/vinext/src/entries/runtime-entry-module.ts
@@ -1,17 +1,6 @@
 import fs from "node:fs";
 import { fileURLToPath } from "node:url";
-
-/**
- * Convert Windows-style backslash path separators to forward slashes.
- *
- * Generated entry modules embed absolute filesystem paths inside `import`
- * statements. On Windows the OS-native paths use `\` which is invalid in JS
- * module specifiers, so every entry generator normalizes paths through this
- * helper before stringifying them into the emitted code.
- */
-export function normalizePathSeparators(p: string): string {
-  return p.replace(/\\/g, "/");
-}
+import { normalizePathSeparators } from "../utils/path.js";
 
 /**
  * Resolve a sibling module path relative to a caller's `import.meta.url`,
